Refactor: Adopting per-module folder unit test structure
- Moved inline unit tests into dedicated tests.rs files within each module folder for idiomatic Rust test organization
- Enables clean separation of primary code and unit tests
